+++ src/RadixDlt.NetworkGateway.GatewayApi/Validators/NonFungibleDataRequestValidator.cs
@@ -86,7 +86,7 @@ internal class NonFungibleDataRequestValidator : AbstractValidator<GatewayModel.
         RuleFor(x => x.Cursor)
             .Base64();
 
-        RuleFor(x => x.Limit)
+        RuleFor(x => x.LimitPerPage)
             .GreaterThan(0)
             .LessThanOrEqualTo(endpointOptionsSnapshot.Value.MaxPageSize);
     }