+++ src/RadixDlt.NetworkGateway.GatewayApi/Validators/NonFungibleIdsRequestValidator.cs
@@ -83,7 +83,7 @@ internal class NonFungibleIdsRequestValidator : AbstractValidator<GatewayModel.N
         RuleFor(x => x.Cursor)
             .Base64();
 
-        RuleFor(x => x.Limit)
+        RuleFor(x => x.LimitPerPage)
             .GreaterThan(0)
             .LessThanOrEqualTo(endpointOptionsSnapshot.Value.MaxPageSize);
     }