+++ src/RadixDlt.NetworkGateway.GatewayApi/Validators/StreamTransactionsRequestValidator.cs
@@ -82,7 +82,7 @@ internal class StreamTransactionsRequestValidator : AbstractValidator<GatewayMod
         RuleFor(x => x.Cursor)
             .Base64();
 
-        RuleFor(x => x.Limit)
+        RuleFor(x => x.LimitPerPage)
             .GreaterThan(0)
             .LessThanOrEqualTo(endpointOptionsSnapshot.Value.MaxPageSize);
     }