+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 _Release Date: ???_
 
 - `/transaction/recent` endpoint replaced with `/stream/transactions` returning user and non-user transactions.
+- `limit` request parameter has been renamed to `limit_per_page`. 
 
 ### 0.1.1 - Betanet V2
 _Release Date: 14.02.2022_